Update shell scripts for Buzz to bring inline with other devices
Updated shell scripts for Buzz, so that the way they are laid out resembles
how other devices are.  Removed old copy-files.sh script which is also
no-longer needed (remnants of old CM6 build scripts)

There was also a slight issue with setup-makefiles.sh which was still causing
the old LiveWallpapers to be pulled in (from change I9b0de2a3) as they were
still cropping up and being included in the latest nightly's.

Change-Id: Ic33f1885c096981b42818b7242dc5280c3b5e429