buzz: Enable offmode charging
At first I tried to use the binary files from original Wildfire rom, but
it didn't work correctly with them. Phone would reboot if pulling usb plug
while in recovery, notification light would be lit when in recovery etc.
So i tried with the ones from the liberty, and it seems to work surprisingly
well with them. Credits to Ladios for them.

A custom recovery kernel is needed as the regular one is too big, making
the image exceed the limit.

Change-Id: I48511f587a3f135412904623143dfc21b0451161